updated grade crud

+++ database/migrations/2023_01_24_212256_create_grades_table.php
@@ -16,8 +16,8 @@ class CreateGradesTable extends Migration
         Schema::create('grades', function (Blueprint $table) {
             $table->id();
             $table->string('grade');
-            $table->integer('min');
-            $table->integer('max');
+            $table->double('min');
+            $table->double('max');
             $table->unsignedBigInteger('created_by');
             $table->unsignedBigInteger('updated_by');
             $table->softDeletes();
